i think i brought up the version?????

+++ src/main/java/net/szum123321/ariadne_glasses/data/Octreee.java
@@ -1,7 +1,7 @@
 package net.szum123321.ariadne_glasses.data;
 
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3f;
+import net.minecraft.util.math.Vec3d;
 import net.szum123321.ariadne_glasses.VecfUtil;
 
 import java.util.stream.IntStream;
@@ -12,15 +12,15 @@ public interface Octreee {
 
     IntStream query(BlockPos pos, double radius);
 
-    default void insert(Vec3f pos, int id) {
+    default void insert(Vec3d pos, int id) {
         insert(VecfUtil.round(pos), id);
     }
 
-    default IntStream query(Vec3f pos, double radius) {
+    default IntStream query(Vec3d pos, double radius) {
         return query(VecfUtil.round(pos), radius);
     }
 
-    default void remove(Vec3f pos) {
+    default void remove(Vec3d pos) {
         remove(VecfUtil.round(pos));
     }
